[Screen Ruler] Fix edge detection at index 0 and SAD truncation in bounds mode
FindEdge broke out of the loop as soon as the decremented coordinate hit
zero, so it never read the pixel at row/column 0. With the measured region
starting at index 1 the detected boundary came out one pixel too wide. Stop
the scan when the coordinate goes negative instead, so index 0 is tested like
any other position. The incrementing branch was already correct, since there
maxDim is one past the last valid index.

PixelsClose masked the SAD result with 0xFF in total mode, which is the
default path because per-channel comparison is off by default. The four
channel differences can reach 1020, so any sum above 255 wrapped back down
into the tolerance range and made clearly different pixels compare as close.
At the shipped tolerance of 30, white against #A9A9A9 sums to 258 and scored
2, so edge detection walked straight through that boundary. Mask with 0xFFFF,
which holds the maximum without clipping.

Fixes #46947
Fixes #46946

+++ src/modules/MeasureTool/MeasureToolCore/BGRATextureView.h
@@ -118,7 +118,9 @@ struct BGRATextureView
         else
         {
             // Method 2: Test whether sum of all channel differences is smaller than tolerance
-            const int32_t score = _mm_cvtsi128_si32(_mm_sad_epu8(distances, _mm_setzero_si128())) & std::numeric_limits<uint8_t>::max();
+            // The 4 channel differences sum to at most 1020, so masking with 0xFF wrapped
+            // anything above 255 back down into the tolerance range.
+            const int32_t score = _mm_cvtsi128_si32(_mm_sad_epu8(distances, _mm_setzero_si128())) & std::numeric_limits<uint16_t>::max();
             return score <= tolerance;
         }
     }
